Rework type classes again
As it turns out factoring out range check was not good idea. First it's
questionable from performance PoV it relies heavily on GHC to optimize away
result of check. More importantly it breaks down with hermitian matrices
in addition of index transform it requires conjugation of result of

Much more reasonable idea is to assume that index is in range and let each
instance deal with questions of its internal representation.

In the end attempt to provide generic indexing without any checks was
misguided. Those have preconditions specific for each data type so we should
provide monomorphic variant for that purpose.