Introduce BLASInt and LAPACKInt
BLAS could use both 32 and 64 bit integers so we want to make sure that we pass
correctly sized integers. LAPACK however seems to always use 32bit. This makes
it possible to switch to 64bit or use both variants as compile time configuration

Additionally we check during compilation that sizes are correct.